+++ lib/src/constants.dart
@@ -25,8 +25,6 @@ class Bishop {
   static const int invalid = -1;
   static const int hand = -2;
   static const Square empty = 0;
-  // TODO: replace promoFlag with the original piece type
-  static const int promoFlag = 1;
   static const defaultSeed = 7363661891;
 
   static const List<int> playerDirection = [-1, 1, -1, -1];